+++ server/api/allen/users/_userId@string/validators.ts
@@ -2,5 +2,5 @@ import { z } from 'zod'
 import { defineValidators } from './$relay'
 
 export default defineValidators(() => ({
-  params: z.object({ userId: z.string() })
+  params: z.object({ userId: z.string() }),
 }))
