+++ styles/css.ts
@@ -1,6 +1,4 @@
-import styled from 'styled-components'
-
-const Origin = styled.div`
+/*const Origin = styled.div`
   box-sizing: border-box;
   display: block;
   border-collapse: separate;
@@ -196,80 +194,5 @@ export const Vertical = styled(StraitLine)`
 export const CenterBox = styled(Origin)`
   width: 252px;
   height: 216px;
-`
-
-export const MainBox = styled(Origin)`
-  float: left !important;
-`
-
-export const Cell = styled(Origin)`
-  width: 24px;
-  height: 24px;
-  font-size: 10px;
-  line-height: 23px;
-  background-image: url(https://minesweeper.online/img/skins/hd/closed.svg?v=3);
-  float: left;
-  background-size: 100%;
-  font-weight: 700;
-  text-align: center;
-`
-export const CellOpen = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type0.svg?v=3);
-`
-
-export const RedBomber = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/mine_red.svg?v=3);
-`
-
-export const Bomber = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/mine.svg?v=3);
-`
-
-export const Flag = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/flag.svg?v=3);
-`
-
-export const OpenOne = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type1.svg?v=3);
-`
-
-export const OpenTwo = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type2.svg?v=3);
-`
-
-export const OpenThree = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type3.svg?v=3);
-`
-
-export const OpenFour = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type4.svg?v=3);
-`
-
-export const OpenFive = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type5.svg?v=3);
-`
-
-export const OpenSix = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type6.svg?v=3);
-`
-
-export const OpenSeven = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type7.svg?v=3);
-`
-export const OpenEight = styled(Cell)`
-  background-image: url(https://minesweeper.online/img/skins/hd/type8.svg?v=3);
-`
-
-export const Bottom = styled(Origin)`
-  width: 18px;
-  height: 16.5px;
-  background-image: url(https://minesweeper.online/img/skins/hd/corner_bottom_left_wide_2x.png?v=3);
-  float: left;
-`
-export const BottomsBottom = styled(Origin)`
-  width: 18px;
-  height: 16.5px;
-  background-image: url(https://minesweeper.online/img/skins/hd/t_left_2x.png?v=3);
-  float: left;
-  background-size: 100% 100%;
-`
+`*/
+export {}
